refactor: split trait_impl/mod.rs under 500-LoC cap
Extract the DFlash EAGLE ctx-slot append helpers (dflash_accept_append,
dflash_eagle_accept_append, dflash_eagle_kgamma_append) into a sibling
dflash_append.rs, following the _dispatch delegation pattern already
used throughout trait_impl/*.rs.
